Weight is now of the same type as result.

+++ BrandesKernels.cl
@@ -56,7 +56,7 @@ __kernel void vcsr_forward(
     __global int* voff,
     __global int* ptr,
     __global int* adj,
-    __global int* weight,
+    __global float* weight,
     __global int* dist,
     __global int* sigma,
     __global float* delta
@@ -84,7 +84,7 @@ __kernel void vcsr_forward(
         }
       }
       if (my_off == 0) {
-        delta[my_map] = (float) weight[my_map] / my_sigma;
+        delta[my_map] = weight[my_map] / my_sigma;
       }
     }
   }
@@ -147,7 +147,7 @@ __kernel void vcsr_backward_reduce(
 __kernel void vcsr_sum(
     const int global_id_range,
     const int source,
-    __global int* weight,
+    __global float* weight,
     __global int* dist,
     __global int* sigma,
     __global float* delta,
@@ -155,7 +155,7 @@ __kernel void vcsr_sum(
     ) {
   const int my_i = get_global_id(0);
   if (my_i < global_id_range && my_i != source && dist[my_i] != -1) {
-    bc[my_i] += (delta[my_i] * sigma[my_i] - 1) * (float) weight[source];
+    bc[my_i] += (delta[my_i] * sigma[my_i] - 1) * weight[source];
   }
 }
